--- first_try/main.go
@@ -1,26 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"sync"
-)
-
-func printHello() {
-
-	var wg sync.WaitGroup
-
-	wg.Add(1)
-	go func() {
-
-		defer wg.Done()
-
-		fmt.Println("Hello from goroutine!")
-
-	}()
-
-	wg.Wait()
-}
-
-func main() {
-	printHello()
-}
