+++ third_example/third.go
@@ -1,0 +1,40 @@
+// Использование канала для передачи данных
+// Задача: Напишите функцию, которая создает горутину,
+// отправляющую числа от 1 до 5 в канал, а затем в main извлекает их и складывает,
+// результат выводит в консоль.
+
+package main
+
+import (
+	"fmt"
+)
+
+func createGorutines(num int) <-chan int {
+
+	ch := make(chan int)
+
+	go func() {
+
+		defer close(ch)
+
+		for i := 1; i <= num; i++ {
+			ch <- i
+		}
+	}()
+
+	return ch
+}
+
+func main() {
+	var number int
+	fmt.Scan(&number)
+
+	channel := createGorutines(number)
+	sum := 0
+	for val := range channel {
+
+		sum += val
+
+	}
+	fmt.Printf("Sum = %d\n", sum)
+}
